requirements.txt
# ====================== LangChain 大语言模型应用框架 ======================
# LangChain 核心框架，用于构建基于LLM的应用程序（链、代理、记忆等）
langchain>=1.0.7
# LangChain 的 OpenAI 官方集成包，用于调用 OpenAI 的模型（如GPT-4）
langchain-openai>=1.0.3
# 社区贡献的大量第三方集成，包括各种工具、向量库和LLM提供商
langchain-community>=0.4.1
# LangChain 的核心抽象和基础运行时，上述包均依赖于此
langchain-core>=1.0.7

# ====================== Web 服务框架与服务器 ======================
# 高性能、易用的现代 Python Web 框架，用于构建 API 和 Web 应用
fastapi>=0.100.0
# 高性能的 ASGI 服务器，用于运行 FastAPI 应用
uvicorn>=0.23.0

# ====================== 数据库驱动 ======================
# MySQL官方Python驱动，用于连接和操作MySQL数据库
mysql-connector-python~=9.4.0
# 备选的纯Python MySQL驱动，安装更简单。当前项目首选官方驱动，故此行注释。
# pymysql>=1.1.1

# ====================== 向量数据库与云服务 ======================
# Pinecone 向量数据库的客户端库，用于存储和检索向量嵌入（常用于AI应用）
pinecone~=7.3.0
# 阿里云灵积模型平台 SDK，用于调用通义千问等模型
dashscope>=1.14.0

# ====================== 工具与辅助库 ======================
# 用于从 .env 文件加载环境变量，方便配置管理（如API密钥）
python-dotenv>=1.0.0
# 简单优雅的 HTTP 客户端库，用于发送网络请求
requests>=2.31.0
# 基于类型提示的数据验证与设置管理库，FastAPI 的核心依赖之一
pydantic~=2.11.7
# 底层 HTTP 客户端库，requests 依赖于此。通常不需要单独指定，除非有特殊版本需求。
# urllib3~=2.5.0
